+++ test/fixtures/edge/dbo.SampleProc.sql
@@ -1,0 +1,5 @@
+CREATE PROCEDURE dbo.sample_proc_fixture
+AS
+BEGIN
+    CREATE TABLE #temp (id INT PRIMARY KEY);
+END